pyproject.toml
[tool.poetry]
name = "repair-plan-service"
version = "0.0.0"
description = "Repair plan service build with service-platform and deployed with Gekkota"
authors = ["Rick Ducott <rick.ducott@geckorobotics.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
uvicorn = "^0.18.2"
pytest = "^7.1.2"
trio = "^0.21.0"
sniffio = "^1.3.0"
pandas = "^2.0.1"
fastapi = "^0.95.2"
gecko-service-platform = "^0.16.2"
gecko-rolodex-lib = "^0.22.4"
geojson-pydantic = "^0.6.2"
httpx = "^0.24.1"
plotly = "^5.14.1"
reportlab = "3.6.12"
kaleido = "0.2.1"
openpyxl = "^3.1.2"
xlsxwriter = "^3.1.1"

[tool.poetry.dev-dependencies]

[[tool.poetry.source]]
name = "geckorobotics-pypi"
url = "https://geckorobotics.jfrog.io/artifactory/api/pypi/pypi/simple"
default = false
secondary = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
